requirements.txt
Django==4.1.7
djangorestframework==3.14.0
djoser==2.1.0
python-dotenv==1.0.0
Django==4.1.7
djangorestframework==3.14.0
djoser==2.1.0
gunicorn==20.1.0
psycopg2==2.9.5
python-dotenv==1.0.0
requests==2.28.2
